added proper logging console instead of println. added timeout to rpc-reply consumer(publisher). created kotlin-style builder for message properties.

+++ src/main/kotlin/dk/renner/website/config/KotlinConfig.kt
@@ -15,7 +15,7 @@ class KotlinConfig {
     @Scope(ConfigurableBeanFactory.SCOPE_SINGLETON)
     fun rabbitMqConnectionFactory(rabbitMqProperties: RabbitMqProperties) = ConnectionFactory().apply {
         host = rabbitMqProperties.host
-        port = rabbitMqProperties.port.toInt()
+        port = rabbitMqProperties.port
         username = rabbitMqProperties.user
         password = rabbitMqProperties.password
     }
